refactor(packages): name the Mux element flavors for their engines
The hls.js flavor held the unqualified module name while its SPF counterpart carried a
suffix, so the default flavor was identifiable only by what it was not called. Both are now
named for the engine underneath — `hlsjs-media`/`spf-media` in html, `hlsjs`/`spf` in
React — and each element index re-exports whichever flavor is the default. The html define
entries get a directory per element for the same reason.

Packaging follows the layout: explicit `exports` subpaths point the flavor-neutral
specifiers at those index files, so `media/mux-video` keeps resolving to whatever its index
re-exports. The React flavors also need their own build entries — the index-only entry glob
left `spf.tsx` unbuilt entirely, and separate entries are what keep importing one flavor
from pulling in the other engine.
