+++ packages/html/src/define/media/mux-audio/hlsjs.ts
@@ -1,5 +1,5 @@
-import { MuxAudio } from '../../media/mux-audio';
-import { safeDefine } from '../safe-define';
+import { MuxAudio } from '../../../media/mux-audio/hlsjs-media';
+import { safeDefine } from '../../safe-define';
 
 export class MuxAudioElement extends MuxAudio {
   static readonly tagName = 'mux-audio';
