+++ packages/html/src/media/mux-audio/index.ts
@@ -1,1 +1,1 @@
-export * from './media';
+export * from './hlsjs-media';
