--- packages/react/src/media/mux-video/spf/index.ts
@@ -1,1 +1,0 @@
-export * from './media';
